Disasters patch
Fixed the Disasters patch compatibility
Removed incompatible .Net components
Fixed string constant in i18
Fixed Dutch translation
