+++ TimeWarpMod/SunController.cs
@@ -3,10 +3,8 @@ using ColossalFramework.IO;
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Text;
-using System.Threading.Tasks;
 using UnityEngine;
 
 namespace TimeWarpMod
@@ -32,6 +30,10 @@ namespace TimeWarpMod
 
         SimulationManager sim = Singleton<SimulationManager>.instance;
         private uint dayOffsetFrames;
+
+        public void EarlyUpdateData()
+        {
+	}
 
         public void Awake()
         {
